requirements.txt
# Core dependencies
watchdog>=3.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
loguru>=0.7.0
aiofiles>=23.0.0
jinja2>=3.1.0
python-multipart>=0.0.6

# mnamer - media file renaming
mnamer>=2.5.0

# Additional utilities
python-dateutil>=2.8.0
